pref(user): 添加auth服务，优化token校验逻辑
