+++ pkg/discovery/client.go
@@ -10,6 +10,23 @@ var (
 	baseUrl = "consul://%s/%s?wait=14s&healthy=true"
 )
 
+func NewGrpcClient(name string, addr string) (*grpc.ClientConn, error) {
+	var grpcOptions = []grpc.DialOption{grpc.WithInsecure()}
+
+	addr = fmt.Sprintf(baseUrl, addr, name)
+	grpcOptions = append(grpcOptions, grpc.WithDefaultServiceConfig(`{"loadBalancingPolicy": "round_robin"}`))
+
+	conn, err := grpc.Dial(
+		addr,
+		grpcOptions...,
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	return conn, nil
+}
+
 func NewBalanceGrpcClient(ac *config.AppConfig) (map[string]*grpc.ClientConn, error) {
 	servers := map[string]*grpc.ClientConn{}
 	for _, sc := range ac.Discovers {
